still seeding very slowly, even with indexes taken out
