+++ server/postgres/copyScript.sql
@@ -1,3 +1,29 @@
-COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/workspaceLocations.sql' WITH DELIMITER AS '|';
+COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/workspaceLocations1.sql' WITH DELIMITER AS '|';
+COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/locationPointers1.sql' WITH DELIMITER AS '|';
 
-COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/locationPointers.sql' WITH DELIMITER AS '|';
+COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/workspaceLocations2.sql' WITH DELIMITER AS '|';
+COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/locationPointers2.sql' WITH DELIMITER AS '|';
+
+-- COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/workspaceLocations3.sql' WITH DELIMITER AS '|';
+-- COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/locationPointers3.sql' WITH DELIMITER AS '|';
+
+-- COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/workspaceLocations4.sql' WITH DELIMITER AS '|';
+-- COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/locationPointers4.sql' WITH DELIMITER AS '|';
+
+-- COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/workspaceLocations5.sql' WITH DELIMITER AS '|';
+-- COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/locationPointers5.sql' WITH DELIMITER AS '|';
+
+-- COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/workspaceLocations6.sql' WITH DELIMITER AS '|';
+-- COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/locationPointers6.sql' WITH DELIMITER AS '|';
+
+-- COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/workspaceLocations7.sql' WITH DELIMITER AS '|';
+-- COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/locationPointers7.sql' WITH DELIMITER AS '|';
+
+-- COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/workspaceLocations8.sql' WITH DELIMITER AS '|';
+-- COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/locationPointers8.sql' WITH DELIMITER AS '|';
+
+-- COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/workspaceLocations9.sql' WITH DELIMITER AS '|';
+-- COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/locationPointers9.sql' WITH DELIMITER AS '|';
+
+-- COPY public."WorkspaceLocations" ("uuid","workspaceSlug","workspaceId","rawAddress","formattedAddress","streetName","streetNumber","neighborhood","city","state","country","countryCode","zipCode") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/workspaceLocations10.sql' WITH DELIMITER AS '|';
+-- COPY public."LocationPointers" ("uuid","workspaceId","longitude","latitude","workspaceLocation_id") FROM '/Users/munzmoses/RPT25/nearby-workspaces/server/postgres/dataFiles/locationPointers10.sql' WITH DELIMITER AS '|';
