+++ server/postgres/createDb.sql
@@ -1,4 +1,2 @@
 DROP DATABASE workspacelocations;
 CREATE DATABASE workspacelocations;
-
--- psql postgres < schema.sql
