+++ server/postgres/index.js
@@ -8,7 +8,7 @@ const password = '';
 const sequelize = new Sequelize(database, user, password, {
   host,
   dialect: 'postgres',
-  // logging: false,
+  logging: false,
 });
 
 sequelize.authenticate()
