--- server/postgres/locationPointers.sql
@@ -1,10 +1,0 @@
-1c76b6f2-42ec-4335-afc2-abcba49bb70c|1|169.3526|71.6854|4324a5dd-915a-44f1-820f-a165f16ad8e4
-21464957-5597-4703-a3f1-3da989d20926|2|5.9547|-19.0626|d833e75b-4d76-4ff7-88c0-c18b9b9fd128
-f478838b-b3ff-4995-8ebf-83750db671b8|3|97.264|33.6810|9db63d0e-392e-47ed-9f35-ce7e0b6ccc45
-df3a4d68-f71a-4476-8a7a-460e2ed3e7e5|4|116.6348|-9.7722|3e74e38b-3cab-48eb-a178-30bb4f59f6b5
-0a11b836-5efc-42cb-b3b0-497dcb86e3ad|5|66.7426|87.9578|9a468d26-570e-450c-ba83-884d98cf405c
-b4a52470-8d18-4a31-911e-539e23939791|6|131.4694|-67.9203|824da8be-88b1-4193-90c0-d0ffe367ed61
-4f294e97-4135-4dc4-9658-40f215e8659b|7|170.875|-14.8029|f048ee92-c9cc-432b-a6c9-293fb9bae96e
-3a22c03e-741e-46d9-9886-99e6c82376da|8|3.7854|65.3225|bfbe3c54-8da3-4195-a479-a6ac677cb014
-4eadbfb2-917c-437c-9e8b-588c2c583bc3|9|85.2|57.1908|b37834c2-62bd-4686-b91a-25f3a473aab1
-fbf1f7fe-a335-4395-9010-e153e525810f|10|114.401|-42.5438|906d9812-3ce4-4484-ae33-b08a1d159cb3
