+++ server/postgres/pgSeeder.js
@@ -1,70 +1,59 @@
 const faker = require('faker');
 const sequelize = require('./');
-const { WorkspaceLocation, LocationPointer } = require('./models');
 require('./relationship');
+const uuid = require('uuid').v4;
+const fs = require('fs').promises;
+const argv = require('yargs').argv;
+const lines = argv.lines || 100;
 
-
-const saveSeed = async (data, transaction) => {
-  const { uuid } = await WorkspaceLocation.create({ ...data }, { transaction });
-  await LocationPointer.create({
-    longitude: Math.abs(faker.address.longitude()),
-    latitude: faker.address.latitude(),
-    workspaceId: data.workspaceId,
-    workspaceLocation_id: uuid,
-  }, { transaction });
-  if (data.workspaceId === 2) {
-    await transaction.commit();
-    console.log('seeded');
-    process.exit();
-  }
-}
 
 sequelize.sync({ force: true })
   .then(() => {
     console.log('MODELS CONNECTED');
     (async() => {
-        try {
-          const transaction = await sequelize.transaction({autocommit: false});
-          let seeder = 1;
-          for(let i = 0; i < 2; i++) {
-          const streetNumber = faker.random.number(1000);
-          const streetName = faker.address.streetName();
-          const city = faker.address.city();
-          const state = faker.address.stateAbbr();
-          const zipCode = faker.address.zipCode().slice(0, 5);
-          const neighborhood = `${faker.address.direction()} ${city}`;
-          const rawAddress = `${streetNumber} ${streetName}, ${city}, ${state} ${zipCode}, USA`
-          const dataFormat = {
-            workspaceSlug: faker.lorem.slug(),
-            workspaceId: seeder,
-            rawAddress,
-            formattedAddress: rawAddress,
-            streetNumber,
-            streetName,
-            neighborhood,
-            city,
-            state,
-            country: 'United States',
-            countryCode: 'US',
-            zipCode,
-          };
-          await saveSeed(dataFormat, transaction).catch(async(err) => {
-            console.log(err);
-            process.exit();
-          });
-          seeder++;
-          }
-        } catch(err) {
-          console.log(err);
+      let workspaceLocations = ''
+      let locationPointers = '';
+      let previousFileIndex = 1;
+      let currentFileIndex = 1;
+      fs.writeFile(`./dataFiles/workspaceLocations${currentFileIndex}.sql`, workspaceLocations);
+      fs.writeFile(`./dataFiles/locationPointers${currentFileIndex}.sql`, locationPointers);
+      for(let i = 1; i <= lines; i++) {
+        if(currentFileIndex > previousFileIndex) {
+          console.log(currentFileIndex, previousFileIndex);
+          fs.writeFile(`./dataFiles/workspaceLocations${currentFileIndex}.sql`, workspaceLocations);
+          fs.writeFile(`./dataFiles/locationPointers${currentFileIndex}.sql`, locationPointers);
+          previousFileIndex = currentFileIndex;
         }
+        const streetNumber = faker.random.number(1000);
+        const streetName = faker.address.streetName();
+        const city = faker.address.city();
+        const state = faker.address.stateAbbr();
+        const zipCode = faker.address.zipCode().slice(0, 5);
+        const neighborhood = `${faker.address.direction()} ${city}`;
+        const rawAddress = `${streetNumber} ${streetName}, ${city}, ${state} ${zipCode}, USA`
+        const workspaceLocationsUuid = uuid();
+        workspaceLocations += `${workspaceLocationsUuid}|${faker.lorem.slug()}|${i}|${rawAddress}|${rawAddress}|${streetName}|${streetNumber}|${neighborhood}|${city}|${state}|United States|US|${zipCode}\n`
+        locationPointers += `${uuid()}|${i}|${faker.address.longitude(-117, -82)}|${faker.address.latitude(33, 41)}|${workspaceLocationsUuid}\n`;
+        try {
+          await fs.appendFile(`./dataFiles/workspaceLocations${currentFileIndex}.sql`, workspaceLocations);
+          await fs.appendFile(`./dataFiles/locationPointers${currentFileIndex}.sql`, locationPointers);
+          workspaceLocations = '';
+          locationPointers = '';
 
+        } catch(err) {
+          console.log('Error writing seed script file', err);
+          process.exit();
+        }
+        //create a new file for every million
+        if(i.toString().slice(-6) === '000000') {
+          currentFileIndex++;
+        }
+      }
+      console.log('Seeding Complete');
+      process.exit();
     })();
   })
   .catch((err) => {
     console.log(err, 'MODELS NOT SYNCED');
     process.exit();
   });
-
-
-
-// process.exit();
