--- server/postgres/workspaceLocations.sql
@@ -1,10 +1,0 @@
-4324a5dd-915a-44f1-820f-a165f16ad8e4|officiis-tenetur-vitae|1|218 Botsford Flats, Gloverville, NV 08536, USA|218 Botsford Flats, Gloverville, NV 08536, USA|Botsford Flats|218|Southeast Gloverville|Gloverville|NV|United States|US|08536
-d833e75b-4d76-4ff7-88c0-c18b9b9fd128|eum-neque-dolor|2|498 Audreanne Radial, Nelsfurt, WI 34149, USA|498 Audreanne Radial, Nelsfurt, WI 34149, USA|Audreanne Radial|498|West Nelsfurt|Nelsfurt|WI|United States|US|34149
-9db63d0e-392e-47ed-9f35-ce7e0b6ccc45|quia-hic-fugiat|3|406 Bartell Valleys, Ziemannfort, MS 47518, USA|406 Bartell Valleys, Ziemannfort, MS 47518, USA|Bartell Valleys|406|East Ziemannfort|Ziemannfort|MS|United States|US|47518
-3e74e38b-3cab-48eb-a178-30bb4f59f6b5|ut-facilis-sit|4|167 Langworth Drive, West Hubert, ND 25014, USA|167 Langworth Drive, West Hubert, ND 25014, USA|Langworth Drive|167|Northeast West Hubert|West Hubert|ND|United States|US|25014
-9a468d26-570e-450c-ba83-884d98cf405c|rerum-eveniet-sed|5|902 Mante Key, Murrayton, PA 32034, USA|902 Mante Key, Murrayton, PA 32034, USA|Mante Key|902|Northwest Murrayton|Murrayton|PA|United States|US|32034
-824da8be-88b1-4193-90c0-d0ffe367ed61|iusto-eligendi-ut|6|111 Lilliana Park, Lake Kamren, OR 68702, USA|111 Lilliana Park, Lake Kamren, OR 68702, USA|Lilliana Park|111|Southeast Lake Kamren|Lake Kamren|OR|United States|US|68702
-f048ee92-c9cc-432b-a6c9-293fb9bae96e|perspiciatis-rerum-molestiae|7|438 Funk Valley, West Jalon, HI 67322, USA|438 Funk Valley, West Jalon, HI 67322, USA|Funk Valley|438|Southwest West Jalon|West Jalon|HI|United States|US|67322
-bfbe3c54-8da3-4195-a479-a6ac677cb014|iste-qui-provident|8|204 Koss Fall, Lake Kelvinchester, IA 64346, USA|204 Koss Fall, Lake Kelvinchester, IA 64346, USA|Koss Fall|204|Northeast Lake Kelvinchester|Lake Kelvinchester|IA|United States|US|64346
-b37834c2-62bd-4686-b91a-25f3a473aab1|et-non-provident|9|912 Parisian Tunnel, Sabrinaview, CT 13024, USA|912 Parisian Tunnel, Sabrinaview, CT 13024, USA|Parisian Tunnel|912|Southwest Sabrinaview|Sabrinaview|CT|United States|US|13024
-906d9812-3ce4-4484-ae33-b08a1d159cb3|at-suscipit-doloremque|10|806 Howell Plaza, Jamisonstad, NV 46186, USA|806 Howell Plaza, Jamisonstad, NV 46186, USA|Howell Plaza|806|West Jamisonstad|Jamisonstad|NV|United States|US|46186
